remove layout view
